added sub-pixel version
This code only works for the first 637 frames of the video. Which I think is actually enough as the oscillations become minute after about 150-200 frames. So we could use this. Still need to figure out what the problem is. The fourier transforms seem to be bugging, dont understand why yet. This uses a function called phase_correlation which you will also need.